+++ src/Commands/App/Renamer.php
@@ -4,7 +4,6 @@ namespace LaravelZero\Framework\Commands\App;
 
 use Illuminate\Support\Str;
 use LaravelZero\Framework\Commands\Command;
-use Symfony\Component\Console\Input\InputArgument;
 
 /**
  * This is the Laravel Zero Framework renamer command class.
@@ -14,16 +13,12 @@ use Symfony\Component\Console\Input\InputArgument;
 class Renamer extends Command
 {
     /**
-     * The name of the console command.
-     *
-     * @var string
+     * {@inheritdoc}
      */
-    protected $name = 'app:rename';
+    protected $signature = 'app:rename {name? : The new name}';
 
     /**
-     * The console command description.
-     *
-     * @var string
+     * {@inheritdoc}
      */
     protected $description = 'Perform an application rename';
 
@@ -35,14 +30,6 @@ class Renamer extends Command
         $this->alert('Renaming the application...');
 
         $this->rename();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function configure(): void
-    {
-        $this->addArgument('name', InputArgument::OPTIONAL);
     }
 
     /**