+++ src/Contracts/Commands/Component/Installer.php
@@ -2,9 +2,6 @@
 
 namespace LaravelZero\Framework\Contracts\Commands\Component;
 
-use LaravelZero\Framework\Commands\Component\Installer as InstallCommand;
-use LaravelZero\Framework\Contracts\Providers\Composer as ComposerContract;
-
 /**
  * This is the Laravel Zero Framework component install contract.
  *
@@ -13,13 +10,17 @@ use LaravelZero\Framework\Contracts\Providers\Composer as ComposerContract;
 interface Installer
 {
     /**
+     * Returns the component name.
+     *
+     * @return string
+     */
+    public function getComponentName(): string;
+
+    /**
      * Installs the component and returns the result
      * of the installation.
      *
-     * @param \LaravelZero\Framework\Commands\Component\Installer $command
-     * @param \LaravelZero\Framework\Contracts\Providers\Composer $command
-     *
      * @return bool
      */
-    public function install(InstallCommand $command, ComposerContract $composer): bool;
+    public function install(): bool;
 }