+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- Core: All configs are auto detected.
+- Core: Database component installer now publishes a `database.php` config.
+- `config/config.php` removed.
+- `config/app.php` added contains now only app key config.
+
 ## [3.10.0] - 2017-11-01
 ### Changed
 - Uses symfony default command by default.